requirements.txt
playwright>=1.35.0
beautifulsoup4
lxml
httpx
aiofiles
aiosqlite
PyYAML
pandas
trafilatura
langid
datasketch
tiktoken
pyarrow
duckdb
prometheus_client
feedparser
playwright-stealth
